use float64 instead of float32

+++ evaluater.go
@@ -2,5 +2,5 @@ package darwin
 
 // Evaluater computes and set individual Fitness
 type Evaluater interface {
-	Evaluate(Individual) (Fitness float32)
+	Evaluate(Individual) (Fitness float64)
 }
